+++ ModuleDevelop/Test.js
@@ -21,16 +21,23 @@ const server = http.createServer((req, res) => {
 server.listen(port, hostname, () => {
   console.log(`Server running at http://${hostname}:${port}/`);
   
-  var result1 = dhm.login("id1","pw101", function(){
+  var result1 = dhm.login("id1","pw101", (result)=>{
     console.log("실패 결과 출력");
-    console.log(arguments);
+    
+    console.log(result);
   });
   console.log("result1",result1);
 
-  var result2 = dhm.login("id1","pw1", function(){
+  var result2 = dhm.login("id1","pw1", (result)=>{
     console.log("성공 결과 출력");
-    console.log(arguments);
+    console.log(result);
   });
+
+  var join = dhm.join("angus","1234","이무현","poro","angus99@naver.com","010-8355-3460",(result)=>{
+    console.log("성공 결과 출력");
+    console.log(result);
+  });
+
   console.log("result2",result2);
 });
